requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
aiosqlite==0.19.0
slack-sdk==3.26.1
aiofiles==23.2.1
pydub==0.25.1